+++ CHANGELOG.md
@@ -35,6 +35,7 @@ title: Changelog
 * Upgrade from `@web/rollup-plugin-import-meta-assets@1.0.6` to `@web/rollup-plugin-import-meta-assets@1.0.7`.
 * Upgrade from `@web/test-runner@0.13.4` to `@web/test-runner@0.14.0`.
 * Upgrade from `@web/test-runner-mocha@0.7.2` to `@web/test-runner-mocha@0.7.5`.
+* Upgrade from `@custom-elements-manifest/analyzer@0.4.1` to `@custom-elements-manifest/analyzer@0.6.4`.
 
 ## 9.0.0 (2022-07-19)